add Runtime::block_on API
Adds the `Runtime::block_on` and `tokio_threadpool::Sender::block_on`
methods, which take a future and run it to completion on the Tokio
runtime. The future provided to these methods must be `Send`, but not
`'static` (unlike `Runtime::spawn`). If the future panics, the panic is
propagated to the calling thread.

The API is pretty straightforward, however there are two things that
might need tweaking:

1) `Sender::block_on` returns a nested `Result` in order to indicate a
   threadpool error. This is a bit awkward, but I'm not sure what will
   be the preferred way to indicate a failure here.  This isn't a
   problem for `Runtime::block_on`, because the existing convention is
   to panic on threadpool error.

2) the API docs warn not to call `block_on` from an async context.
   This could be enforced by asserting or debug asserting that the
   calling thread is not a tokio-threadpool worker, but that check is
   not included in this commit.